pyproject.toml
[tool.poetry]
name = "code jam 10 qualifier"
version = "0.1.0"
description = "code base of pydiscord code jam 10"
authors = ["NaviTheCoderboi <102732282+NaviTheCoderboi@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.0.0"
numpy = "^1.25.2"
opencv-python = "^4.8.0.76"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
